+++ app/models/recording_studio/concerns/recording_hierarchy.rb
@@ -44,14 +44,11 @@ module RecordingStudio
       def descendants
         return [] if id.blank?
 
-        root_id = RecordingStudio.root_recording_id_for(root_recording_or_self)
-        recordings_by_parent = self.class.unscoped
-                                   .where(root_recording_id: root_id)
-                                   .where.not(id: id)
-                                   .order(updated_at: :desc)
-                                   .group_by(&:parent_recording_id)
+        ids = descendant_id_chain
+        return [] if ids.empty?
 
-        collect_descendants(recordings_by_parent, parent_id: id)
+        recordings_by_id = self.class.unscoped.where(id: ids).index_by(&:id)
+        ids.filter_map { |descendant_id| recordings_by_id[descendant_id] }
       end
 
       def self_and_descendants
@@ -67,22 +64,12 @@ module RecordingStudio
       private
 
       def ancestor_id_chain
-        ancestor_ids = []
-        current_parent_id = parent_recording_id
-
-        while current_parent_id.present?
-          ancestor_ids << current_parent_id
-          current_parent_id = self.class.unscoped.where(id: current_parent_id).pick(:parent_recording_id)
-        end
-
-        ancestor_ids
+        RecordingStudio::Tree.ancestor_ids(parent_recording_id)
       end
 
-      def collect_descendants(recordings_by_parent, parent_id:)
-        children = Array(recordings_by_parent[parent_id])
-        children.flat_map do |child|
-          [child] + collect_descendants(recordings_by_parent, parent_id: child.id)
-        end
+      def descendant_id_chain
+        root_id = RecordingStudio.root_recording_id_for(root_recording_or_self)
+        RecordingStudio::Tree.descendant_ids(id, root_recording_id: root_id)
       end
 
       def assign_root_recording_id
@@ -113,7 +100,6 @@ module RecordingStudio
         errors.add(:parent_recording_id, "must belong to the same root recording")
       end
 
-      # rubocop:disable Metrics/AbcSize, Metrics/CyclomaticComplexity, Metrics/MethodLength
       def parent_recording_must_not_create_cycle
         return if parent_recording_id.nil?
 
@@ -124,22 +110,11 @@ module RecordingStudio
 
         return if id.blank?
 
-        visited_ids = Set.new
-        current_parent_id = parent_recording_id
+        ancestor_ids = RecordingStudio::Tree.ancestor_ids(parent_recording_id)
+        return unless ancestor_ids.map(&:to_s).include?(id.to_s)
 
-        while current_parent_id.present?
-          return if visited_ids.include?(current_parent_id)
-
-          if current_parent_id == id
-            errors.add(:parent_recording_id, "cannot be itself or a descendant recording")
-            return
-          end
-
-          visited_ids << current_parent_id
-          current_parent_id = self.class.unscoped.where(id: current_parent_id).pick(:parent_recording_id)
-        end
+        errors.add(:parent_recording_id, "cannot be itself or a descendant recording")
       end
-      # rubocop:enable Metrics/AbcSize, Metrics/CyclomaticComplexity, Metrics/MethodLength
     end
   end
 end